+++ lib/raxx.ex
@@ -670,7 +670,7 @@ defmodule Raxx do
   end
 
   defp redirect_page(url) do
-    html = html_escape(url)
+    html = EEx.HTML.escape_to_binary(url)
     "<html><body>This resource has moved <a href=\"#{html}\">here</a>.</body></html>"
   end
 
@@ -756,84 +756,6 @@ defmodule Raxx do
       false ->
         {:error, "module `#{Macro.to_string(module)}` is not available."}
     end
-  end
-
-  ######## COPIED FROM PLUG ########
-
-  @doc ~S"""
-  Escapes the given HTML to string.
-
-      iex> html_escape("foo")
-      "foo"
-
-      iex> html_escape("<foo>")
-      "&lt;foo&gt;"
-
-      iex> html_escape("quotes: \" & \'")
-      "quotes: &quot; &amp; &#39;"
-  """
-  @spec html_escape(String.t()) :: String.t()
-  def html_escape(data) when is_binary(data) do
-    IO.iodata_to_binary(to_iodata(data, 0, data, []))
-  end
-
-  @doc ~S"""
-  Escapes the given HTML to iodata.
-
-      iex> html_escape_to_iodata("foo")
-      "foo"
-
-      iex> html_escape_to_iodata("<foo>")
-      [[[] | "&lt;"], "foo" | "&gt;"]
-
-      iex> html_escape_to_iodata("quotes: \" & \'")
-      [[[[], "quotes: " | "&quot;"], " " | "&amp;"], " " | "&#39;"]
-
-  """
-  @spec html_escape_to_iodata(String.t()) :: iodata
-  def html_escape_to_iodata(data) when is_binary(data) do
-    to_iodata(data, 0, data, [])
-  end
-
-  escapes = [
-    {?<, "&lt;"},
-    {?>, "&gt;"},
-    {?&, "&amp;"},
-    {?", "&quot;"},
-    {?', "&#39;"}
-  ]
-
-  for {match, insert} <- escapes do
-    defp to_iodata(<<unquote(match), rest::bits>>, skip, original, acc) do
-      to_iodata(rest, skip + 1, original, [acc | unquote(insert)])
-    end
-  end
-
-  defp to_iodata(<<_char, rest::bits>>, skip, original, acc) do
-    to_iodata(rest, skip, original, acc, 1)
-  end
-
-  defp to_iodata(<<>>, _skip, _original, acc) do
-    acc
-  end
-
-  for {match, insert} <- escapes do
-    defp to_iodata(<<unquote(match), rest::bits>>, skip, original, acc, len) do
-      part = binary_part(original, skip, len)
-      to_iodata(rest, skip + len + 1, original, [acc, part | unquote(insert)])
-    end
-  end
-
-  defp to_iodata(<<_char, rest::bits>>, skip, original, acc, len) do
-    to_iodata(rest, skip, original, acc, len + 1)
-  end
-
-  defp to_iodata(<<>>, 0, original, _acc, _len) do
-    original
-  end
-
-  defp to_iodata(<<>>, skip, original, acc, len) do
-    [acc | binary_part(original, skip, len)]
   end
 end