+++ test/eex/html_engine_test.exs
@@ -1,4 +1,0 @@
-defmodule EEx.HTMLEngineTest do
-  use ExUnit.Case
-  doctest EEx.HTMLEngine
-end