+++ CHANGELOG.md
@@ -16,6 +16,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `Raxx.set_body/2` will raise an exception for responses that cannot have a body.
 - `Raxx.set_body/2` automatically adds the "content-length" if it is able.
 
+### Removed
+
+- `Raxx.html_escape/1` replaced by `EEx.HTML.escape_to_binary/1`.
+
 ## [0.15.11](https://github.com/CrowdHailer/raxx/tree/0.15.11) - 2018-09-04
 
 ### Added